+++ x/cash/msg.go
@@ -12,24 +12,21 @@ func init() {
 	migration.MustRegister(1, &UpdateConfigurationMsg{}, migration.NoModification)
 }
 
-// Ensure we implement the Msg interface
-var _ weave.Msg = (*SendMsg)(nil)
-
 const (
-	pathSendMsg                      = "cash/send"
-	pathUpdateConfigurationMsg       = "cash/update_configuration"
-	sendTxCost                 int64 = 100
+	sendTxCost int64 = 100
 
 	maxMemoSize int = 128
 	maxRefSize  int = 64
 )
 
-// Path returns the routing path for this message
+var _ weave.Msg = (*SendMsg)(nil)
+
+// Path returns the routing path for this message.
 func (SendMsg) Path() string {
-	return pathSendMsg
+	return "cash/send"
 }
 
-// Validate makes sure that this is sensible
+// Validate makes sure that this is sensible.
 func (s *SendMsg) Validate() error {
 	var err error
 	if coin.IsEmpty(s.Amount) || !s.Amount.IsPositive() {
@@ -65,8 +62,7 @@ func (s *SendMsg) DefaultSource(addr []byte) *SendMsg {
 	}
 }
 
-// FeeTx exposes information about the fees that
-// should be paid
+// FeeTx exposes information about the fees that should be paid.
 type FeeTx interface {
 	GetFees() *FeeInfo
 }
@@ -130,5 +126,5 @@ func (m *UpdateConfigurationMsg) Validate() error {
 }
 
 func (*UpdateConfigurationMsg) Path() string {
-	return pathUpdateConfigurationMsg
+	return "cash/update_configuration"
 }