+++ x/sigs/msg.go
@@ -1,6 +1,7 @@
 package sigs
 
 import (
+	"github.com/iov-one/weave"
 	"github.com/iov-one/weave/errors"
 	"github.com/iov-one/weave/migration"
 )
@@ -10,11 +11,11 @@ func init() {
 }
 
 const (
-	pathBumpSequenceMsg = "sigs/bump_sequence"
-
 	maxSequenceIncrement = 1000
 	minSequenceIncrement = 1
 )
+
+var _ weave.Msg = (*BumpSequenceMsg)(nil)
 
 func (msg *BumpSequenceMsg) Validate() error {
 	if err := msg.Metadata.Validate(); err != nil {
@@ -30,5 +31,5 @@ func (msg *BumpSequenceMsg) Validate() error {
 }
 
 func (BumpSequenceMsg) Path() string {
-	return pathBumpSequenceMsg
+	return "sigs/bump_sequence"
 }