+++ x/validators/msg.go
@@ -11,14 +11,11 @@ func init() {
 	migration.MustRegister(1, &ApplyDiffMsg{}, migration.NoModification)
 }
 
-// Ensure we implement the Msg interface
 var _ weave.Msg = (*ApplyDiffMsg)(nil)
 
-const pathApplyDiffMsg = "validators/apply_diff"
-
-// Path returns the routing path for this message
+// Path implements weave.Msg interface.
 func (*ApplyDiffMsg) Path() string {
-	return pathApplyDiffMsg
+	return "validators/apply_diff"
 }
 
 func (m *ApplyDiffMsg) Validate() error {