+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## HEAD
 
 - `bnsd/x/username` genesis initializer implemented and included in `bnsd`.
-- Support gov proposal vote, deletion and tally in `bnscli` 
+- Support gov proposal vote, deletion and tally in `bnscli`
 - Added `x/utils.ActionTagger`: all `bnsd` transactions now have
   `action=${msg.Path()}` tags. If there is a batch, there is one tag per
   sub-message. If it is a governance tally, the TallyMsg as well as the
@@ -12,6 +12,8 @@
 Breaking changes
 
 - Unify all message paths to follow pattern `<package>/<message_name>`
+- `app.Router` interface was changed. Handler registration requires a message
+  and not message path.
 
 ## 0.17.0